requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
email-validator==2.1.1
Werkzeug==3.0.4
itsdangerous==2.2.0
Jinja2==3.1.4
SQLAlchemy==2.0.34
python-dotenv==1.0.1

Flask-Mail==0.9.1  
bcrypt==4.2.0 


Flask-Login==0.6.3
